+++ entrypoints/sherpa-romeo.content/SherpaRomeoButton.tsx
@@ -1,8 +1,7 @@
-import { Accordion, AccordionDetails, AccordionSummary, Button, Stack } from "@mui/material";
-import { ThemeProvider } from "@mui/material";
-import { themeLight } from "@/utils/theme";
 import PublisherPolicyData from "@/components/SherpaRomeo/PublisherPolicyData";
 import { PublisherPolicy } from "@/types";
+import { themeLight } from "@/utils/theme";
+import { Button, ThemeProvider } from "@mui/material";
 import { useState } from "react";
 import getPublisherPolicies from "./common";
 export default function SherpaRomeoButton({ issn }: { issn: string }) {
@@ -14,10 +13,12 @@ export default function SherpaRomeoButton({ issn }: { issn: string }) {
     }
     return (
         <ThemeProvider theme={themeLight}>
-            <div className="col-xs-12" >
+            <div className="col-xs-12 needs-xs-spacing" >
                 <Button variant="contained" onClick={onClick}>Buscar en SherpaRomeo</Button>
             </div>
             <br />
+            <br />
+
 
             <PublisherPolicyData PublisherPolicies={publisherPolicies} />
         </ThemeProvider>